Final changes.

+++ src/main/java/by/epam/pialetskialiaksei/command/admin/subject/DeleteSubjectCommand.java
@@ -36,29 +36,29 @@ public class DeleteSubjectCommand implements Command {
                     + subjectsIdJson);
             int[] subjectsIdArray = new Gson().fromJson(subjectsIdJson, int[].class);
             List<Integer> subjectsId = new ArrayList<>();
-            for(int subId: subjectsIdArray){
+            for (int subId : subjectsIdArray) {
                 subjectsId.add(subId);
             }
-            List<Integer>subjectsIdThatCantBeDeleted = new ArrayList<>();
+            List<Integer> subjectsIdThatCantBeDeleted = new ArrayList<>();
             MarkDAO markDAO = new MarkDAO();
-            for(Integer subjectId: subjectsId){
+            for (Integer subjectId : subjectsId) {
                 Mark mark = markDAO.findEntrantMark(subjectId);
-                if(mark == null){
+                if (mark != null) {
                     subjectsIdThatCantBeDeleted.add(subjectId);
                 }
             }
             SubjectDAO subjectDAO = new SubjectDAO();
-            if(subjectsIdThatCantBeDeleted.isEmpty()) {
+            if (subjectsIdThatCantBeDeleted.isEmpty()) {
                 subjectDAO.delete(subjectsId);
-            }else{
+            } else {
                 subjectsId.removeAll(subjectsIdThatCantBeDeleted);
-                subjectDAO.delete(subjectsId);
-                //                result = "{\"errorEng\":\"you cant delete subject, because it is used.\", \"errorRu\":\"\\u0412\u044b \u043d\u0435 \u043c\u043e\u0436\u0435\u0442\u0435 \u0443\u0434\u0430\u043b\u0438\u0442\u044c \u043f\u0440\u0435\u0434\u043c\u0435\u0442, \u0442\u0430\u043a \u043a\u0430\u043a \u043e\u043d \u0438\u0441\u043f\u043e\u043b\u044c\u0437\u0443\u0435\u0442\u0441\u044f.\"}";
-//                result = "{\"errorEng\":\"you cant delete subject, because it is used.\", \"errorRu\":\"Вы не можете удалить предмет, так как он уже используется.\"}";
-                result = "{\"errorEng\":\"You cant delete subjects with ids:" +
+                if (!subjectsId.isEmpty()) {
+                    subjectDAO.delete(subjectsId);
+                }
+                result = "{\"errorEng\":\"You cant delete subjects with ids: " +
                         subjectsIdThatCantBeDeleted.toString().replaceAll(REGEX, "") +
                         ", because they are already used.\"," +
-                        "\"errorRu\":\"Вы не можете удалить предметы с id:" +
+                        "\"errorRu\":\"Вы не можете удалить предметы с id: " +
                         subjectsIdThatCantBeDeleted.toString().replaceAll(REGEX, "") +
                         ", так как они уже используется\"}";
             }
